transaction_driver: simplify wiring of transactions
Previously, we were using a counter to determine whether both
prepare and replicate had finished. Now, in the leader, we don't
even trigger replicate until prepare has finished. So, that counter
was just confusing things, and should be removed.

I also changed the wiring a bit so that we don't use the future callback
to trigger the replicate phase of the transaction. Instead, it's just tacked
onto the end of the prepare code, which makes the code a little more
"straight-line" and easier to follow (in my opinion).

I also updated the replica transaction code to use separate booleans
to track what part of the transaction was still pending (leader commit
or local prepare). I think this is again a little easier to read than
the counter. The structure there is also modified to look more like
the structure of the leader.

It's possible that something is now broken around transaction abort,
but all of the tests appear to pass. So, if there is a bug, we should
add more unit tests as well.

Change-Id: I1e7bafed8a4d987a914ab1d321cf3dc48b5875db
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3980
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>