cfile: remove file size from Open() variant of CFileReader
When creating a CFileReader, it's strange to get the file size as input
when the file itself supports fetching it via Size(). I thought perhaps some
callers were "customizing" the file size, but no, they're all just fetching
the size directly from the underlying file.

Changing this means deferring initialization of the file size from the
constructor to Init() (since Size() may fail), but it's cleaner. And it's
marginally more performant when callers used Env::GetFileSize(), as we avoid
a (hashed) path lookup by switching from stat() to fstat().

Change-Id: I7cfc968c023cfc5af4c316361ea85bbfac2d9a48
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3985
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins