KUDU-405 Part 5. Decoupled RowProjector and ExecutionEngine.
This commit decouples the row projector from the execution and the
row projector from the generated functions. This is less of a cleanup commit
and more of a "setup" commit for the code cache diff.

The decoupling is necessary for several reasons, the most important of which
is that the projection code for projection functions applies to a wider
set of schema pairs than the sole pair that is assigned to a row projector.

Additionally, the decoupling also makes the jitted code lifetime explicitly
maintained by the JITCodeOwner, which will improve clarity when multiple
classes rely on generated code from the same module.

In order to ensure that its code is valid, the row projector takes
shared ownership of the JITCodeOwner during construction.

Change-Id: Iae6c629b4121f2ecc698ebd367aef767cb1325bf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3997
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins