KUDU-422: Determine permanent_uuid if not specified.
If a QuorumPeerPB doesn't have a permanent_uuid set, determine it
through an added RPC call which retrieves NodeInstancePB for the peer.
The reason for this change is to allow specifiying master quorums on the
command line.

TODO (another revision): move UpdatePermamentUuids() to a more appropriate
place, e.g., raft_consensus.

Change-Id: Iea44767a1f338abf9952416c6871f0e4195acad4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3988
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: jenkins