test_result_server: add a simple list_failed_tests endpoint
This produces results equivalent to determine_flaky_tests.py --list-tests-only,
but is much faster, since it's backed by the test results database instead of
by Jenkins' Enterprise XML Repository Of Doom. In fact, this returns results in
about 183ms whereas in a recent test run, Jenkins took about 20 minutes (a
whopping 6557x speedup!)

I tested this patch facebook-style. That is to say, I deployed in production
and crossed my fingers that it would work. You can see the results at:

http://a1228.halxg.cloudera.com:8080/list_failed_tests?num_days=3&build_pattern=%25kudu-test%25

Longer term it would be nice to switch over the determine_flaky_tests.py script
to use something like this endpoint, but that was more work than I wanted to
put in, tonight, and instead just went for the targeted fix in build-and-test.sh

Change-Id: I1b5aabb0029e7e5c3afa7b2c1561444a121f0033
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3991
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>