Fix up a lot of confusion between column IDs and column indexes
This fixes one of many issues related to alter-schema: we are very messy about
when we pass around column *ids* and when we pass around column *indexes*. This
patch changes Schema and SchemaBuilder, in DEBUG mode only, to start assigning
column IDs at 10 instead of 0.  This ensures that, in most cases, the column ID
and column index are not equal, and triggered several bugs where we'd been
conflating them.

In order to fix the issues that popped up, this patch makes a number of changes:

* AddColumnUpdate always takes an id, never an index. Historically, we used to
  use RowChangeLists on the client and sometimes encode indexes, and sometimes IDs.
  Nowadays, it's always server-side, and it doesn't make sense to maintain this
  behavior.

* Simplified ProjectUpdate a bit - it had some stub code for "alter type"
  functionality which we have no near-term plans to implement. Currently
  it is just confusing in the code, so I made it more straightforward using
  Schema functions directly.

* ApplyRowUpdate is updated to properly use column IDs instead of column indexes.
  This was one place where we were relying on column IDs and indexes being
  one-to-one.

* Fixed an incorrect DCHECK in Schema::column_by_id which asserted that ids were
  always lower than the number of columns.

* Fixed Schema::find_column_by_id() to not allow being called on a schema
  without IDs, since it would probably indicate a bug.

* Updated a number of unit tests to explicitly lookup column IDs in the schema
  instead of assuming that they are 0-based.

It's surprising that more tests didn't fail, because I'm pretty sure there are
more lurking issues where IDs and indexes are conflated. I'll continue to work
through these in follow-on patches.

Change-Id: I31b2700f79c79af1569f42284424891352caeed2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4010
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>