Add a DEBUG-mode schema compatibility check to row projector.
This commit explicitly defines "schema compatibility," which is the condition
necessary for two pairs of (base, projection) schemas to have the same
row projecting code.

This definition is implemented in a compatibility-checking function which
is utilized in DEBUG mode to check that the schemas which a row projector is
given and the schemas from which its codegen functions were built are the same.

Change-Id: Ibecf25d6d3baef58c17bd3c654fe9aaae1dfd1ab
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4001
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins