KUDU-410 - Start proper ChangeConfigTransactions on startup
We're currently special-casing (and kind-of hacking) consensus start-up.
Thats a problem as we would have a different path for start up than we
would have for subsequent configuration changes. Moreover we really didn't
have the actual config-change-after-start path so this patch is a progress
in both directions.

This makes the intial configuration, provided at startup, function as
an election result with implicit votes from peers (as in we must record
those votes when we have the means to do so). So the potential leader
now starts as candidate, promotes itself to leader through a new BecomeLeader()
method and initiates a leader side transaction that will try to push that
configuration to other peers. Only if/when a majority accepts the message does
the leader peer start to accept writes and proceed with the remainder
of the startup.

This also means that we no longer block waiting for the intitial config
round to be accepted on startup, something that was filling up worker
queues before.

Follow ups:
- static leader change on a quiet quorum tests

Change-Id: I826bd5b137e602b5a2d3582e323bb21089f75a46
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4162
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins