test: make check for no leftover test output more robust
This took way too much time, so here's a far too long commit message.

The problem: when a known flaky test fails, is retried, and passes, the test
output from the flaky run is left behind, causing the "no leftover test
output when all tests pass" check to fail. There are two distinct causes:
1. A test fails in ASSERT and ~KuduTest purposefully retains the test output
   for what is presumed to be further debugging.
2. A test fails in CHECK and aborts immediately. No cleanup is done.

The first fix I tried was adding a 'never' option to --test_leave_files. I
expected this to address #1 but not #2. Turns out it's not a great solution
to #1 either, because not all tests use test_util.cc and I didn't think it
made sense to define it globally, putting it in tests that wouldn't use it.
Nonetheless, I retained the change because I think it's generally useful.

Next, I tried using an atexit() handler to clean up outstanding test output.
This didn't work for #2 either: abort() doesn't invoke atexit() handlers,
and I didn't feel comfortable changing the glog fail function to exit().

Next, I changed run-test.sh to compare the contents of TEST_TMPDIR before
and after the test runs, and to delete the difference. This was a promising
approach because it handles both #1 and #2 in the same way. But it doesn't
work when tests are run in parallel, as many of them do. So in the end I
settled on a variation: include argv[0] in the test directory name and match
on that prefix in run-test.sh.

Change-Id: I933193a7eb6674d91c8aa9a75394931c8e66fce8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4067
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>