KUDU-490, KUDU-497: Replicate CatalogManager.
This revision implements distributed consensus for CatalogManager.

Currently the quorum configuration is passed in on the command line:

For the leader, the configuration options are:
--leader # To specify that the node is a leader.
--follower_addresses=host1:port,host2:port,host3:port.

For the followers, the configuration options are:
--leader_address=leader_host:port # The leader node.
--follower_addresses=host1:port,host2:port # List of all hosts other
                                           # than the current host
                                           # and the leader.

NOTE: currently, hot fail over is not supported. To change the
current master, first all masters must be restarted with the new
configuration and then the tablet servers must be pointed at the
new master; this is not safe as we don't do a real leader
election when picking a master as per the raft algorithm (which states
that we must pick a master that has a majority of commits).

Change-Id: I5a1678f2bd636b2458b9c0b7f331f925d5481cc3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4190
Tested-by: jenkins
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>