When submitting a write to tablet that is not ready yet reply with the appropriate error
When submitting a write to a tablet fails before any task is triggered we're swallowing the
error and just issuing a warning. This is problematic since the client will not get a response
if the prepare task failed to submit at all. Moreover we should reply with a particular code
if the tablet is not in RUNNING state so the client can try again later.

This patch captures the Status from tablet_peer_->SubmitWrite(), in TabletServiceImpl::Write,
and makes sure the client gets an appropriate response if the write could not be submitted.

Change-Id: Ib164ead386de8e2387bc28220b4186de30036cb0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4085
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>