Add a (disabled) test that makes sure that we get TABLET_NOT_RUNNING when bootstrap fails
I was following Todd's suggestion on another patch to corrupt a tablet's data on
purpose for it to fail bootstrap, so that in turn we could reveal a bug in the tablet
server where it would not reply to the client. The specific bug I was hoping to reveal didn't
exist but on the process of doing this I uncovered another bug where we segfault on shutdown
if the tablet did not bootstrap successfully.

This patch adds a test that verifies that we return the right thing when a tablet fails to
bootstrap and that also tests that we're able to shutdown a tablet in such a state. It's
disabled as this is fixed in follow up patches.

Change-Id: I7ee74d8128b6484e751290ec584b2e41b4733cfc
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4103
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>