Decouple row projector from code cache.
This commit cleans up the nested JITPayload class and re-brands it as
a JITSchemaPair, while moving some responsibilities which should be
dependent on the JITSchemaPair to the class. Previously, the methods
relating to schema pair compatibility, schema pair encoding, etc., were
all distributed amongst varying classes that shouldn't really be in
charge of those methods (like the CodeCache and the RowProjector).

More importantly, this decoupling sets the ground for future
expansion of classes to codegen.

Change-Id: I6db63a180880a9cabab7a372d73f0833699b848f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4116
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>