Add a consensus specific error PB
We will have to have a bunch of consensus-specific error codes that
peers can need to set on the response status so that the sender peer
can take specific action. Things like replying INVALID_TERM: set by
replicas when the received message has a term lower than the last
one they know about. Moreover this should sit in the consensus proto
since likely it's going to be used by both the tserver and the master.

Change-Id: I689eb2fb79d143f31f3cf0a5dea237cd494f8923
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4112
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>