Removed race in metrics for CompilationManager at shutdown.
Running client-test's ClientTest.TestReplicatedMultiTabletTableFailover
may reveal* a race between compilation manager, which relies on the tablet
server's metric registry to log its info, and the tablet server's destructor.

The registry would get deleted (in case of failover), but the manager
would still be accepting incoming requests, which would get logged
in deleted counters.

To resolve this, we opted with the ThreadMgr approach (defined in
thread.cc), where a server would "StartInstrumentation" with the
compilation manager, creating FunctionGuages which observe the
the CompilationManager's metrics passively without requiring any
CompilationManager dependency on a given server's registry.

*apparently, TSAN doesn't always catch this. Here's a link to a build where
it did:
http://sandbox.jenkins.cloudera.com/job/kudu-gerrit/4739/BUILD_TYPE=TSAN,label=kudu-gerrit-slaves/

Change-Id: Ic7ba5394c45e30beb524f9668d51523391fcf63b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4118
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins