Redesigned codegen directory
To facilitate extension, this commit gets rid of the aggregate class
which is held as the cache value and instead creates a JITWrapper interface
(which is code + pointers to functions), where ref-counted pointers to
JITWrappers are the values in cache.

Using an enum, each derived class can ensure a unique key header for its type,
so using the same cache is acceptable for all codegen needs.

This will facilitate use of codegen in other contexts (wheras before codegen
could only be used in the schema-pair-projection context, since each
projection pair would map to the same place).

Change-Id: I34e1883b2ba079eeb6f5b68e6bba24afde9505ad
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4139
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>