Config. Change: Send old, new quorum to TabletPeer and fix visibility issue
We need to send the old and the new quorum to the tablet peer when there is
a configuration change so that it can take action when the roles change and
update the tablet manager accordingly.

We also had a visibility bug where we were letting the tablet peer know of
the state change before the commit was durable. This was revealed as problem
on a follow-up patch, when we use the tablet peer's state changed method
to mark it as running when the log is yet to observe the commit message go
through. Because we were changing the tablet to running before the message
was in the log but after a transaction was executed, certain anchoring
related tests in tablet_peer-test would fail.

This patch fixes both by making the ChangeConfigTransactionState store the
'old' state, after Prepare() completes. It then proceeds to call
TabletPeer::ConsensusStateChanged() with both the old and the new quorum
but only when the transaction's commit message is durable, i.e. on Finish().

Change-Id: I936e15b530f702d9ccabd52e23f1bd96c4a48ff4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4154
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>