fs: Modify data block open calls to enable sync-on-close
This ensures that we have less likelihood of missing an fsync() call.
We also fsync() parent directories as soon as we open the block files.

This will change again once we have the block manager rewrite committed,
but for now just trying to close durability gaps.

Command-line option coming in follow-up patch.

Change-Id: I134c1bb9bf7caf6c8119eb256473ff1fc2388984
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4231
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>