[java client] Refactor how we pass common configurations to MR jobs
There are a few configurations that almost all MR jobs we ship will accept, like the master's
address or the operation timeout. This patch puts them all behind a utility called
CommandLineParser that can be used to easily get the master's address from the Configuration
as long as it was processed by the GenericOptionsParser.

Change-Id: I5c84bd926c705a1b9ef78c3172ae5a6946b0734a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4229
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>