remote bootstrap: Add flag to tablet metadata to detect remote bootstrap failure
When remote bootstrap crashes in the middle, we don't want local
bootstrap to be successful at the next startup, otherwise we could end
up running with a corrupted tablet. We can detect such issues with a
flag that lives in the tablet metadata, which will be integrated with
the rest of remote bootstrap.

Added a unit test to ensure tablet bootstrap fails when the metadata
flag is set to COPYING.

Change-Id: I28b294fd63b1ba6dec46610519a958ba9b2c3c1e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4233
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>