remote bootstrap: Update in-memory state when replacing superblock
This fixes a bug in the ReplaceSuperBlock implementation that was
originally overlooked. The TabletMetadata has to update its own
in-memory state when the superblock is replaced by the remote copy
during tablet bootstrap.

This was caught by remote_bootstrap-itest, which fails if we don't do
this since it thinks that the metadata is still stuck in
REMOTE_BOOTSTRAP_COPYING state even after we replace the superblock with
the DONE version.

Change-Id: Id4824dfcefd61ef604ff39e713e9ce62670956a5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4234
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>