metadata: make TabletMetadata locking more fine-grained
This is meant to prevent an issue JD was seeing on a test cluster
trying to create a table with lots of tablets.

The issue seemed to be contention between several threads:

T1) some txn execution thread:
      <various fsync calls>
      TabletMetadata::Flush()  <holding TabletMetadata::lock_>
      ChangeConfigTransaction::Apply()
T2) heartbeat thread
      <blocked on TabletMetadata::lock_>
      kudu::metadata::TabletMetadata::Quorum
      TSTabletManager::CreateReportedTabletPB
      TSTabletManager::GenerateIncrementalTabletReport <holding TSTabletManager::lock_>
T*) any other threads trying to create new tablets
      <blocked on TSTabletManager::lock_ write mode>

The solution here is to split the TabletMetadata lock into one which
protects just the data, and another which protects just the flushing process.
This means that slow IO doesn't prevent other threads from reading. It
_does_ mean that other threads can now read un-flushed data from the
metadata object, but that should be OK -- we always write things to the WAL
before we update the metadata, guaranteeing that it's durable.

Change-Id: I6f44d8f2bd3a59de1208ef93d9e049d1399f2b37
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4291
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins