[java client] Re-connect scanners when the TS disconnects
Before this patch the scanners were throwing an error as soon as the TS they were
communicating with got disconnected, as a way to enforce scanner correctness. The
problem is that the servers disconnect automatically after 10 seconds (this client
doesn't have a keep-alive mechanism) so, for example, a Mapper that scans Kudu
that has to spill can easily get this disconnection.

This patch adds a re-connect when we see that the TS is disconnected.

Change-Id: I128111455e51bee2f76b7d5b4479bd3df03145a0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4314
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>