Refactor, fix and document RaftConsensus::UpdateReplica
We had a few bugs on this method, for instance we just returned a bad
status when we couldn't write to the WAL, in this scenario we could
attempt to write the same operation multiple times.

This patch refactors the method hollistically but the main point is
that it adds documentation on what should happen, in what order, and
why.

Change-Id: Idbff5001dd02f90be569833dc4876f66ec4dd255
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4306
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>