Add an inner TrackedPeer struct to consensus queue
This adds a simple TrackedPeer struct to the consensus queue so that
we can keep track of information related to a peer that goes beyond
just ConsensusStatusPB's fields (stuff like terms and the last time we
heard from a peer which will be added in follow up patches).

This also helps in printing all the info related to a peer through
its own ToString() method.

Change-Id: Ia929825b389215a47196ac2e2fce095c50337b69
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4344
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>