KUDU-468 [java client] Deadlock between removing a server and adding metadata about it
The fix removes the locking of TabletClient in RemoteTablet.addTabletClient since it was
unnecessary. It also detects if the race is happening in which case it will retry to connect
the TabletClient. Finally it changed the order in removeClientFromCache since it was removing
from ip2client last while newClient is adding to it first.

This is a deadlock that also seems to be in asynchbase but, since it uses a mechanism to keep
the connection alive, they aren't as prone to this race condition.

A deterministic unit test would be almost impossible to write but the big linked list test was
always able to trigger it. Everything worked once this patch was applied.

Change-Id: I5676c107517488b9056b0320cf6b1e9f91195810
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4386
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>