fs: sync groups of blocks during merge compaction
With the BlockManager integrated, I wanted to make use of block group
syncing, a feature we didn't have prior to the BlockManager. My goals were
two-fold: prove that it can actually improve "real world" performance, and
find a reasonable way to plumb it into the code.

I targeted merge compaction because it's localized: all we need is for
RollingDiskRowSetWriter to keep track of opened blocks as it writes them,
and sync them all at the end. That yields (m rowsets * (n cols + 1 bloom
+ 0/1 composite index)) blocks in one group. Plumbing-wise I thought it'd
be cleanest to continue the single-owner block business and retrieve blocks
from CFileWriters when we're done with them.

In addition, Todd suggested that we'd get a more even I/O pattern if we
kicked off asynchronous data flushes as we finish each block (instead of as
one large group at the end). That's in here too.

To control the behavior we use two gflags:
- block_coalesce_sync: whether to sync blocks as a group or individually.
- cfile_flush_block_on_finish: whether to asynchronously flush blocks as we
  finish them.

I ran 'compaction-test --gtest_filter=*BenchmarkMergeWithoutOverlap* \
  --merge_benchmark_num_rowsets=20 --enable_block_data_fsync \
  --cfile_flush_block_on_finish={true,false} \
  --block_coalesce_sync={true,false}' to test the differences. The timings
from the interesting combinations are below:
- cfile_flush_block_on_finish: ~15.5s
- !cfile_flush_block_on_finish and block_coalesce_sync: ~16.5s
- !cfile_flush_block_on_finish and !block_coalesce_sync: ~31s

Change-Id: I0bee2442d0ef5f602791bb96b65496e363a11d9e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4395
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>