cache: clarify lifetime of the keys used in a cache
When reviewing one of Vlad's outstanding patches, I thought I saw a bug where
we weren't properly keeping around the key slices that were put into the cache.
Upon further investigation of the Cache implementation, I realized that it
copies the key, so Vlad's code was fine.

This patch just clarifies the comment on the Insert method.

Change-Id: Ib979a097f43c6491eab0f02f055f1ff2d82ab4c6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4401
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>