Integrate LogReader as a segment tracker in Log and refactor GC
This pulls out segment tracking from the Log and into LogReader. A Log
now maintains a LogReader as a private field and updates it as it rolls
segments, so that reader always has access to all the segments, including
the one being written to.

This also refactors GC to simply use the new methods in LogReader, making
it simpler and cleaner.

Change-Id: Id00170f7c5e9641ad996143dac6041b3859cf285
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4459
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins