fs: rejigger enable_data_block_fsync
Tying enable_data_block_fsync to sync-on-close makes it harder for code
outside of this module to know whether it needs to sync blocks or if they'll
do it themselves. However, if we always use sync-on-close and push
enable_data_block_fsync down into the block manager, it's not as bad.

Of course, I also want it to apply consistently to both data and metadata
block writes, so I did some more plumbing.

Change-Id: Ie83d2b7727c2e581c05f9009a4fcf7f6d03df50d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4482
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins