In remote bootstrap, use sequence number as the key to segments instead of the first OpId
Removing ids from commit operations opens the possiblity for the absence
of operations with ids in segments, so the approach that remote boostrap
was using of keying segments by the id of the first operation no longer
works.

To support this we need to key segments by their sequence number instead.

Change-Id: Icc3b3b0b8ec19e724745c44c256c7ecf216b6156
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4492
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>