Add a footer to log segments
Currently we lazy-write the header when operations are actually appended
to the log as we need to set the segment's initial OpId in the header. This
is problematic for a few reasons, the main of which being that when
we have operations without ids we run into trouble writing them to the
log if they are the first ones in a segment. Another problem, albeit
less significant, is that we do even more extra work when we roll the
log, meaning we increase the latency of the first operations in the new
segment unecessarily.

This patch addresses this by adding a footer where the first operation
in the segment will be written. In fact the footer can take a set
of operations, acting as an index over the segment, something that
will be needed when we need to serve operations to peers from the log.

The footer can be rebuilt from scanning the segment, were we to crash
before having a chance to write it. ReadableLogSegment now has a method
for that.

Finally this moves more logic into WritableLogSegment, similarly to
what Mike did for ReadableLogSegment a while ago.

Change-Id: I4067f78af56541c16a5ba7a4a608ceaacfdeee84
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4497
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins