Make LogReader track current segments and add methods to fetch particular sequences
Indexing log segments now takes two data structures: an ordered map so that we
can find a given segment by OpId, and a vector so that we keep all segments in
sequence, even if they have no operations with ids. Log currently both handles
writes and tracks the readable segments increasing its complexity unecessarily.

This patch makes LogReader the entity responsible for both querying the
readable segment index and keeping the segment sequence, and for keeping
track of the currently available readable segments.

Follow up patches integrate the LogReader in Log and use the new query
methods to perform GC in a simpler way.

Change-Id: I429459d04734584552865525e38f6151fbb557a8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4507
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>