thirdparty: patch gperftools to fix tcmalloc lock contention
This patches gperftools to change the granularity with which objects are moved
between thread-level free lists and the central free list. This improves
performance about 4.5x in a cluster benchmark that JD has been running, and
also in a microbenchmark:

  ./build/latest/full_stack-insert-scan-test \
    --concurrent_inserts=80 \
    --rows_per_batch=5 \
    --inserts_per_client=10000 \
    --gtest_filter=\*MRSOnly\*

Before:
  359.35user 82.49system 1:05.15elapsed 678%CPU (0avgtext+0avgdata 966384maxresident)k
  0inputs+273248outputs (10major+27965minor)pagefaults 0swaps

After:
  51.19user 18.99system 0:12.35elapsed 568%CPU (0avgtext+0avgdata 907216maxresident)k
  0inputs+273232outputs (10major+23146minor)pagefaults 0swaps

After discussion with the current gperftools maintainer, we think this is
happening because, in a workload with a lot of threads, each thread's thread
cache ends up limited to be pretty small. For example, with a 16MB total thread
cache limit (the default) and 200 threads, each thread cache is only 80KB. With
the old default value for TCMALLOC_TRANSFER_NUM_OBJ, we'd need to accumulate
64kb of objects in a given size class before using the efficient bulk transfer
mechanism. Given that this is a substantial portion of the per-thread cache
size, it's unlikely we'd ever see enough data in one class to hit this
threshold, and we never use the fast path.

One option is to tune up the total thread cache size at runtime, but that
comes at the expense of memory, and it's not clear what a "good value"
would be that would prevent the issue.

This fix helps the high-concurrency situation immensely, and even in the
case of only 10 concurrent writers, it provided a small (5-10% boost).

Change-Id: Ic7678c78b5d4133d0ebbbd607b3ba76fda4c7bc8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4510
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins