Add a dump of the compaction policy log to the tablet rowset layout page
In JD's cluster tests, we got in some state where none of the tablets
wanted to compact themselves. It would have been nice to get the verbose
output of the compaction policy for these tablets. This commit adds
this log info to the rowset layout web page.

I tested locally on a tserver with a bit of data in it.

Change-Id: I5e6e54b87625b1a39a86981b3517732f206294b5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4522
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>