fs: remove Sync() and SYNCED state from writable block
We want to enforce the following write workflow:
- Create the block
- Append data to it a bunch of times
- Flush it asynchronously (optional)
- Synchronize it

As such, we needn't differentiate between SYNCED and CLOSED; both states are
effectively the same because no subsequent operations are allowed in either.

So let's remove Sync() and SYNCED. Close() will implicitly synchronize to
disk, obviating the need for the sync_on_close option.

Change-Id: I3c8798c84e05a1d511b5f4c10530077f771edfc0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4525
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>