KUDU-492 Pt 1: Initialize client with multiple masters.
Part one of client support for distributed masters:

1) Allow specifying multiple masters in KuduClientBuilder.

2) Use the RPCs introduced in KUDU-501 to discover the leader
master (with a maximum timeout, default 15 seconds).

3) Modify master_replication-itest to use the new APIs.

TODO (in future diffs):

1) When the current leader master dies, wait (up to the specified
timeout) for a new leader to become discoverable. Test that all
of the tablet lookup operations function correctly with the newly
elected leader.

2) Document and test a scenario where leader master begins a
CreateTable (or other long-running operation), dies in the middle
of the operation, and a new leader is discovered by the
clients.

Change-Id: I99b7d0262bcc3715b14d69515b7e358e185c4e17
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4833
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins