maintenance_manager: fix perf_improvement to be a double
For compaction operations, we are calculating performance improvement as a
double value, usually between 0.0 and a small number (typically <10). However,
within the maintenance manager we were implicitly casting to uint64_t based on
some early assumption that we'd use some scale from 0 to MAX_LONG.

In cluster testing, we saw that "mature" tablets which had no incoming write
workload quickly became "reasonably compacted" but not fully compacted. In this
state, most tablets had a compaction quality between 0.0 and 1.0, which got
truncated down to 0 by the implicit cast. This caused the maintenance manager
to stop compacting the tablets entirely.

Fixing this, we now see that when the write workload lets up, we do continue
to compact tablets.

Change-Id: I1e202fbd2557fcf37f2576c0483856268ac1135c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4561
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins