Fix a CHECK failure in tablet rowset layout web page
The compaction selection code path clamps all rowset sizes to be at least 1MB,
since we use an integer knapsack algorithm, and we don't want to consider <1MB
rowsets to be "free" to compact.

Previously, we were only doing this clamping in the actual compaction policy
code, and not in the code which we use to display the web page showing tablet
layout. This was fine when we didn't run the policy for this page. But, now we
actually run the policy as well, in order to highlight the next compaction that
we want to run. So, we need to clamp to 1MB in all cases.

Change-Id: Ic351cbab9dd58b313c3ca8b4700636efd276e97e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4562
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins