KUDU-459. Make resettable_heartbeater-test less flaky
This refactors the test a little, but primarily jacks up the heartbeat
period on the flaky test (TestResetHeartbeats) to be 800ms, and reduces
the sleep time in the test in that loop to be 20ms before Reset() calls.
Barring some crazy pause for an entire second of the test threads
getting no CPU time, this should reduce or eliminate the flakiness of
this test.

Verified relative non-flakiness by running this test at full niceness
for 100 iterations while running 24 instances of perl -e 'while(1){}' on
my laptop. Seems OK.

I also reduced the number of expected heartbeats from 3 to 2 just to
keep the test time down to a reasonable level.

Change-Id: I61e60e6ee0fcb4816f21d83ea45e62fadc184566
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4548
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>